- Updates the S3 storage provider to wait for a space to be available before adding metadata to it; also allows spaces to be deleted even when the space metadata can't be found - Updates the storeclient tests to check and wait after creating or deleting a space to allow the underlying provider to become consistent
git-svn-id: https://svn.duraspace.org/duracloud-priv/trunk@460 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ storeclient/src/test/java/org/duracloud/client/TestContentStore.java
@@ -53,7 +53,7 @@ public class TestContentStore
 
     static {
         String random = String.valueOf(new Random().nextInt(99999));
-        spaceId = "storeclient-test-space" + random;
+        spaceId = "storeclient-test-space-" + random;
     }
 
     @Override
@@ -93,12 +93,17 @@ public class TestContentStore
     @After
     public void tearDown() throws Exception {
         // Make sure the space is deleted
-        try {
-            for(String spaceId : spaces) {
+        for(String spaceId : spaces) {
+            if(spaceExists(spaceId)) {
                 store.deleteSpace(spaceId);
             }
-        } catch(ContentStoreException cse) {
-            // Ignore, the space was likely already removed
+
+            int maxLoops = 10;
+            for (int loops = 0;
+                 spaceExists(spaceId) && loops < maxLoops;
+                 loops++) {
+                Thread.sleep(1000);
+            }
         }
     }
 
@@ -182,8 +187,27 @@ public class TestContentStore
 
     private void createSpace(String spaceId, Map<String, String> spaceMetadata)
         throws Exception {
-        store.createSpace(spaceId, spaceMetadata);
+        if(!spaceExists(spaceId)) {
+            store.createSpace(spaceId, spaceMetadata);           
+
+            int maxLoops = 10;
+            for (int loops = 0;
+                 !spaceExists(spaceId) && loops < maxLoops;
+                 loops++) {
+                Thread.sleep(1000);
+            }
+        }
+
         spaces.add(spaceId);
+    }
+
+    private boolean spaceExists(String spaceId) throws Exception {
+        try {
+            store.getSpaceMetadata(spaceId);
+            return true;
+        } catch (NotFoundException e) {
+            return false;
+        }
     }
 
     @Test
